Fix: GCC unaligned pointer warnings
The release of GCC 9 added the following warning:

	-Waddress-of-packed-member, enabled by default, warns about an
	unaligned pointer value from the address of a packed member of a
	struct or union.

The warning is triggered in some place in LTTng-UST in cases where we
pass a pointer to get a result. Rather than passing the pointer directly
from the struct member, we get the result into a local storage, then
write into in the struct.

Signed-off-by: Gabriel-Andrew Pollo-Guilbert <gabriel.pollo-guilbert@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>