doc: reformat long lines in doc/examples/Makefile.am
Format the long lines in the all-local target a bit like the "cmake"
target is formatted already.  I think it helps readability to have one
argument per line instead of very long lines.

At the same time, I removed the "cd .." at the end of parentheses.  The
parentheses start a new subshell, so it's unnecessary to do "cd .."
before the subshell exits.

Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>