Cleanup: apply `include-what-you-use` guideline for `size_t`
I saw that some files use the `size_t` type but there are not including
a header for it. The `size_t` type is defined by the following headers:
  <stddef.h>
  <stdio.h>
  <stdlib.h>
  <string.h>
  <time.h>
  <wchar.h>

So, to follow the best practice of including what is used in a file, I
added `#include <stddef.h>` in files using `size_t` but did not include
any of the previously listed header.

Signed-off-by: Francis Deslauriers <francis.deslauriers@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Change-Id: Ica1d82691335294decf13ffcdf4257e35d6a44c2