Cleanup: apply `include-what-you-use` guideline for `fpos_t`
I saw that some files use the `fpos_t` type but there are not including
a header for it. The `fpos_t` type is defined by the following header:
  <stdio.h>

So, to follow the best practice of including what is used in a file, I
added `#include <stdio.h>` in files using `fpos_t`.

Signed-off-by: Francis Deslauriers <francis.deslauriers@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Change-Id: I808ad0b23f04389fd56f89aa001095b771a327d6