Cleanup: apply `include-what-you-use` guideline for `mbstate_t`
I saw that some files use the `mbstate_t` type but there are not including
a header for it. The `mbstate_t` type is defined by the following header:
  <wchar.h>

So, to follow the best practice of including what is used in a file, I
added `#include <wchar.h>` in files using `mbstate_t`.

Signed-off-by: Francis Deslauriers <francis.deslauriers@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Change-Id: I103d144ede1a8488d2662d3b8ba5337accafda99