Cleanup: silence unused parameter `ps` warning
This function was copied here from one of the BSDs by this commit:
e5bc3b0f4d6c0407492ebdea863483925393e1bc with the /*ARGSUSED*/
lint-style comment to prevent any `unused` warning.

It seems clang doesn't support those lint-style comments, so it prints
this warning:
  CC       libustsnprintf_la-core.lo
  mbrtowc_sb.c:39:68: error: unused parameter 'ps' [-Werror,-Wunused-parameter]
  ust_safe_mbrtowc(wchar_t *pwc, const char *s, size_t n, mbstate_t *ps)

So mark that parameter as unused to prevent any warning.

We could also simply remove the unused parameter and change the
signature of the function.

Signed-off-by: Francis Deslauriers <francis.deslauriers@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Change-Id: I09f4fd3938058ed985c1525d8b9b06a27529691e