Cleanup: apply `include-what-you-use` guideline for `uint*_t`
I saw that some files use `uint*_t` types but there are not including a
header for it. The `uint*_t` types is defined by the following header:
  <stdint.h>

So, to follow the best practice of including what is used in a file, I
added `#include <stdint.h>` in files using `uint*_t`.

Signed-off-by: Francis Deslauriers <francis.deslauriers@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Change-Id: I5da13744858a57fc8c9bf6a6cf1b29299c2211cc