Cleanup: apply `include-what-you-use` guideline for `close()`
I saw that some files use the `close()` function but there are not
including a header for it. The `close()` function is defined by the
following header: <unistd.h>

So, to follow the best practice of including what is used in a file, I
added `#include <unistd.h>` in files using `close()`.

Signed-off-by: Francis Deslauriers <francis.deslauriers@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Change-Id: I07e104e957857c869576af7c2704e98584ecc763