fix: undefined symbols for tracepoints in lgpl source
Move the new internal 'urcu-ust' urcu flavor symbols to a separate
library named 'liblttng-ust-common.so' and link 'liblttng-ust.so' and
'liblttng-ust-tracepoint.so' on it.

Add this common library to the pkgconfig file which should handle the
transition for client code using it in its build system.

LGPL_SOURCE client code which doesn't use pkgconfig in its build system
will need to link with '-llttng-ust-common' instead of '-lurcu-bp' like
prior ust versions.

Change-Id: I807f997a2a94cb4f8d1bdf72002c1ae1b9fedf8d
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>