Fix: notifier: use store-release/load-acquire for error counter
The "record_error" operation is executed concurrently with setting the
error counter in the notifier group without locking, so we need to
explicitly provide existance guarantees.

The only visible transition is from NULL -> !NULL, because the only
situation reverting the error counter back to NULL is on destruction of
the notification group, after an RCU synchronisation guarantees that no
record_error can observe this pointer anymore.

Use the stronger "cmm_smp_mb()" to provide store-release/load-acquire
semantics because the explicit store-release and load-acquire are not
implemented in the CMM model yet.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Change-Id: I3bbeecc00ff33c5d8ad1e3fd41c8a0dc866f1943