Hide private usterr-signal-safe.h symbols
The major SONAME bump to '1' gives us the opportunity to hide private
symbols that should never have been visible.

Changing 'ust_loglevel' from a public weak symbol to a private one,
means that each library that uses ust_err will now have it's own private
copy of the loglevel and will have to call 'ust_err_init' in it's
constructor.

While we're at it, namespace the symbols that were previously part of
the ABI.

Change-Id: I031d4ebfaa43fa4fb3c9f60dc6b557b1d30eca18
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>