Move internal elf macros to a private header
The public 'lttng/ust-elf.h' header exposed a dependency on the libelf
headers even though elfutils is only an optional dependency. None of the
macros and static inline functions using the libelf API need to be
publicly available, move them to a private header.

Also move non-namespaced convenience macros to a private header.

Change-Id: I3b75a3a984ffc2a2827664a76c45c7a8fd7b1fdf
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>