Add 'ctf_unused' tracepoint field type
Some tracepoint definitions have one or more arguments that don't have a
corresponding field to allow passing a value to a probe without
recording it in the trace buffer nor using it in the filter.

This results in tracepoint macros generating inline functions with unused
arguments which prevents us from turning on -Wunuse-parameter.

Add a new tracepoint field type name 'ctf_unused' to adress this and
properly handle unused values in the tracepoint macros.

[ Mathieu: ctf_unused() takes only the _src parameter, no "type",
  because an unused "field" does not have any representation in the ring
  buffer nor the filter. If this ends up limiting us in the future, I
  accept all the blame. ]

Change-Id: I56d8024926340e9c965ca01fa24a0fff7b7c5baa
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>