Add 'src' dir to global include path
The objective is to move internal headers from the global 'include'
directory to their respective private directories under 'src'.

Move the private generated 'config.h' to 'src/common/config.h', this way
it can be include with 'common/config.h' and reduce the risk of clashing
with a similarly named file.

This is part of an effort to standardize our autotools setup across
projects to simplify maintenance.

Change-Id: Ie62a81e46458bffb6d31af66411865f23cf1cb4f
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>